Simplify launch tests, move some to lite
